separate special functions from generated c++ file
